demos/xray/xray-event-container/requirements.txt
tensorflow==2.0.*
boto3==1.9.*
pillow==6.2.*
cloudevents==0.2.*
